requirements.txt
dulwich==0.15.0
PyGithub==1.29
PyYAML==3.12
requests==2.9.1
GitPython==2.1.0
Pint==0.7.2
